Add compatibility with cmake build system

+++ controlsoftware/include/serial.h
@@ -20,7 +20,7 @@ struct daqParsed {
 	uint32_t timestamp;
 	int32_t TC_uaf;
 	uint8_t HALL_methane, HALL_LOX;
-	float BATT_voltage;
+	uint16_t BATT_voltage;
 	int16_t PT_methane, PT_LOX, PT_helium, PT_chamber, PT_heliumReg;
 };
